Refactor vault creation
